move print project tree to code context
