+++ mentat/code_context.py
@@ -4,10 +4,12 @@ import os
 from pathlib import Path
 from typing import Dict, Iterable
 
+from termcolor import cprint
+
 from .code_file import CodeFile
 from .config_manager import ConfigManager
 from .errors import UserError
-from .git_handler import get_non_gitignored_files
+from .git_handler import get_non_gitignored_files, get_paths_with_git_diffs
 
 
 def _is_file_text_encoded(file_path):
@@ -50,11 +52,79 @@ def _abs_files_from_list(paths: Iterable[str], check_for_text: bool = True):
     return files_direct, files_from_dirs
 
 
-def _abs_file_paths_from_list(paths: Iterable[str], check_for_text: bool = True):
-    files_direct, files_from_dirs = _abs_files_from_list(paths, check_for_text)
-    return set(map(lambda f: f.path, files_direct)), set(
-        map(lambda f: f.path, files_from_dirs)
+def _get_files(
+    config: ConfigManager, paths: Iterable[str], exclude_paths: Iterable[str]
+) -> Dict[Path, CodeFile]:
+    excluded_files_direct, excluded_files_from_dirs = _abs_files_from_list(
+        exclude_paths, check_for_text=False
     )
+    excluded_files, excluded_files_from_dir = set(
+        map(lambda f: f.path, excluded_files_direct)
+    ), set(map(lambda f: f.path, excluded_files_from_dirs))
+
+    glob_excluded_files = set(
+        os.path.join(config.git_root, file)
+        for glob_path in config.file_exclude_glob_list()
+        # If the user puts a / at the beginning, it will try to look in root directory
+        for file in glob.glob(
+            pathname=glob_path,
+            root_dir=config.git_root,
+            recursive=True,
+        )
+    )
+    files_direct, files_from_dirs = _abs_files_from_list(paths, check_for_text=True)
+
+    # config glob excluded files only apply to files added from directories
+    files_from_dirs = [
+        file
+        for file in files_from_dirs
+        if str(file.path.resolve()) not in glob_excluded_files
+    ]
+
+    files_direct.update(files_from_dirs)
+
+    files = {}
+    for file in files_direct:
+        if file.path not in excluded_files | excluded_files_from_dir:
+            files[file.path] = file
+
+    return files
+
+
+def _build_path_tree(files: Iterable[CodeFile], git_root):
+    tree = {}
+    for file in files:
+        path = os.path.relpath(file.path, git_root)
+        parts = Path(path).parts
+        current_level = tree
+        for part in parts:
+            if part not in current_level:
+                current_level[part] = {}
+            current_level = current_level[part]
+    return tree
+
+
+def _print_path_tree(tree, changed_files, cur_path, prefix=""):
+    keys = list(tree.keys())
+    for i, key in enumerate(sorted(keys)):
+        if i < len(keys) - 1:
+            new_prefix = prefix + "│   "
+            print(f"{prefix}├── ", end="")
+        else:
+            new_prefix = prefix + "    "
+            print(f"{prefix}└── ", end="")
+
+        cur = cur_path / key
+        star = "* " if cur in changed_files else ""
+        if tree[key]:
+            color = "blue"
+        elif star:
+            color = "green"
+        else:
+            color = None
+        cprint(f"{star}{key}", color)
+        if tree[key]:
+            _print_path_tree(tree[key], changed_files, cur, new_prefix)
 
 
 class CodeContext:
@@ -66,41 +136,11 @@ class CodeContext:
     ):
         self.config = config
 
-        self.files: Dict[Path, CodeFile]
+        self.files = _get_files(self.config, paths, exclude_paths)
 
-        self._set_file_paths(paths, exclude_paths)
-
-    def _set_file_paths(
-        self,
-        paths: Iterable[str],
-        exclude_paths: Iterable[str],
-    ) -> None:
-        excluded_files, excluded_files_from_dir = _abs_file_paths_from_list(
-            exclude_paths, check_for_text=False
+    def print_tree(self):
+        _print_path_tree(
+            _build_path_tree(self.files.values(), self.config.git_root),
+            get_paths_with_git_diffs(self.config.git_root),
+            self.config.git_root,
         )
-
-        glob_excluded_files = set(
-            os.path.join(self.config.git_root, file)
-            for glob_path in self.config.file_exclude_glob_list()
-            # If the user puts a / at the beginning, it will try to look in root directory
-            for file in glob.glob(
-                pathname=glob_path,
-                root_dir=self.config.git_root,
-                recursive=True,
-            )
-        )
-        files_direct, files_from_dirs = _abs_files_from_list(paths, check_for_text=True)
-
-        # config glob excluded files only apply to files added from directories
-        files_from_dirs = [
-            file
-            for file in files_from_dirs
-            if str(file.path.resolve()) not in glob_excluded_files
-        ]
-
-        files_direct.update(files_from_dirs)
-
-        self.files = {}
-        for file in files_direct:
-            if file.path not in excluded_files | excluded_files_from_dir:
-                self.files[file.path] = file
